+++ src/handlers/swap.ts
@@ -376,115 +376,6 @@ export const computeSlippageAdjustedAmounts = (
   return results;
 };
 
-export const executeSwap = async ({
-  chainId,
-  gasLimit,
-  maxFeePerGas,
-  maxPriorityFeePerGas,
-  gasPrice,
-  nonce,
-  tradeDetails,
-  wallet,
-  permit = false,
-  flashbots = false,
-}: {
-  chainId: ChainId;
-  gasLimit: string | number;
-  maxFeePerGas: string;
-  maxPriorityFeePerGas: string;
-  gasPrice: string;
-  nonce?: number;
-  tradeDetails: Quote | null;
-  wallet: Wallet | null;
-  permit: boolean;
-  flashbots: boolean;
-}) => {
-  let walletToUse = wallet;
-  const network = ethereumUtils.getNetworkFromChainId(chainId);
-  let provider;
-
-  // Switch to the flashbots provider if enabled
-  if (flashbots && network === Network.mainnet) {
-    logger.debug('flashbots provider being set on mainnet');
-    provider = await getFlashbotsProvider();
-  } else {
-    logger.debug('normal provider being set', network);
-    provider = await getProviderForNetwork(network);
-  }
-
-  if (!walletToUse) {
-    walletToUse = await loadWallet(undefined, true, provider);
-  } else {
-    walletToUse = new Wallet(walletToUse.privateKey, provider);
-  }
-
-  if (!walletToUse || !tradeDetails) return null;
-
-  const { sellTokenAddress, buyTokenAddress } = tradeDetails;
-  const transactionParams = {
-    gasLimit: toHex(gasLimit) || undefined,
-    // In case it's an L2 with legacy gas price like arbitrum
-    gasPrice: gasPrice || undefined,
-    // EIP-1559 like networks
-    maxFeePerGas: maxFeePerGas || undefined,
-    maxPriorityFeePerGas: maxPriorityFeePerGas || undefined,
-    nonce: nonce ? toHex(nonce) : undefined,
-  };
-
-  // Wrap Eth
-  if (
-    sellTokenAddress === ETH_ADDRESS_AGGREGATORS &&
-    buyTokenAddress === WRAPPED_ASSET[chainId]
-  ) {
-    logger.debug(
-      'wrapping native asset',
-      tradeDetails.buyAmount,
-      walletToUse.address,
-      chainId
-    );
-    return wrapNativeAsset(
-      tradeDetails.buyAmount,
-      walletToUse,
-      chainId,
-      transactionParams
-    );
-    // Unwrap Weth
-  } else if (
-    sellTokenAddress === WRAPPED_ASSET[chainId] &&
-    buyTokenAddress === ETH_ADDRESS_AGGREGATORS
-  ) {
-    logger.debug(
-      'unwrapping native asset',
-      tradeDetails.sellAmount,
-      walletToUse.address,
-      chainId
-    );
-    return unwrapNativeAsset(
-      tradeDetails.sellAmount,
-      walletToUse,
-      chainId,
-      transactionParams
-    );
-    // Swap
-  } else {
-    logger.debug(
-      'FILLQUOTE',
-      tradeDetails,
-      transactionParams,
-      walletToUse.address,
-      permit,
-      chainId
-    );
-    return fillQuote(
-      tradeDetails,
-      transactionParams,
-      walletToUse,
-      permit,
-      chainId
-    );
-  }
-};
-
 export const getTokenForCurrency = (
   currency: Asset,
   chainId: ChainId
