+++ src/hooks/useSwapAdjustedAmounts.ts
@@ -1,7 +1,7 @@
 import { ETH_ADDRESS, Quote } from '@rainbow-me/swaps';
 import lang from 'i18n-js';
 import { useSelector } from 'react-redux';
-import { computeSlippageAdjustedAmounts, Field } from '@/handlers/uniswap';
+import { computeSlippageAdjustedAmounts, Field } from '@/handlers/swap';
 import { AppState } from '@/redux/store';
 import { SwapModalField } from '@/redux/swap';
 import { WETH_ADDRESS, WMATIC_POLYGON_ADDRESS } from '@/references';
