+++ src/raps/actions/swap.ts
@@ -1,5 +1,16 @@
 import { Wallet } from '@ethersproject/wallet';
-import { ChainId } from '@rainbow-me/swaps';
+import {
+  ChainId,
+  CrosschainQuote,
+  ETH_ADDRESS,
+  fillCrosschainQuote,
+  fillQuote,
+  Quote,
+  SwapType,
+  unwrapNativeAsset,
+  wrapNativeAsset,
+  WRAPPED_ASSET,
+} from '@rainbow-me/swaps';
 import { captureException } from '@sentry/react-native';
 import { toLower } from 'lodash';
 import {
@@ -8,8 +19,13 @@ import {
   SwapActionParameters,
 } from '../common';
 import { ProtocolType, TransactionStatus, TransactionType } from '@/entities';
-import { estimateSwapGasLimit, executeSwap } from '@/handlers/uniswap';
-import { isL2Network, toHex } from '@/handlers/web3';
+import { estimateSwapGasLimit } from '@/handlers/swap';
+import {
+  getFlashbotsProvider,
+  getProviderForNetwork,
+  isL2Network,
+  toHex,
+} from '@/handlers/web3';
 import { parseGasParamsForTransaction } from '@/parsers';
 import { additionalDataUpdateL2AssetToWatch } from '@/redux/additionalAssetsData';
 import { dataAddNewTransaction } from '@/redux/data';
@@ -17,8 +33,185 @@ import store from '@/redux/store';
 import { greaterThan } from '@/helpers/utilities';
 import { AllowancesCache, ethereumUtils, gasUtils } from '@/utils';
 import logger from '@/utils/logger';
+import { Network } from '@/helpers';
+import { loadWallet } from '@/model/wallet';
 
 const actionName = 'swap';
+
+export const executeSwap = async ({
+  chainId,
+  gasLimit,
+  maxFeePerGas,
+  maxPriorityFeePerGas,
+  gasPrice,
+  nonce,
+  tradeDetails,
+  wallet,
+  permit = false,
+  flashbots = false,
+}: {
+  chainId: ChainId;
+  gasLimit: string | number;
+  maxFeePerGas: string;
+  maxPriorityFeePerGas: string;
+  gasPrice: string;
+  nonce?: number;
+  tradeDetails: Quote | null;
+  wallet: Wallet | null;
+  permit: boolean;
+  flashbots: boolean;
+}) => {
+  let walletToUse = wallet;
+  const network = ethereumUtils.getNetworkFromChainId(chainId);
+  let provider;
+
+  // Switch to the flashbots provider if enabled
+  if (flashbots && network === Network.mainnet) {
+    logger.debug('flashbots provider being set on mainnet');
+    provider = await getFlashbotsProvider();
+  } else {
+    logger.debug('normal provider being set', network);
+    provider = await getProviderForNetwork(network);
+  }
+
+  if (!walletToUse) {
+    walletToUse = await loadWallet(undefined, true, provider);
+  } else {
+    walletToUse = new Wallet(walletToUse.privateKey, provider);
+  }
+
+  if (!walletToUse || !tradeDetails) return null;
+
+  const { sellTokenAddress, buyTokenAddress } = tradeDetails;
+  const transactionParams = {
+    gasLimit: toHex(gasLimit) || undefined,
+    // In case it's an L2 with legacy gas price like arbitrum
+    gasPrice: gasPrice || undefined,
+    // EIP-1559 like networks
+    maxFeePerGas: maxFeePerGas || undefined,
+    maxPriorityFeePerGas: maxPriorityFeePerGas || undefined,
+    nonce: nonce ? toHex(nonce) : undefined,
+  };
+
+  // Wrap Eth
+  if (
+    sellTokenAddress === ETH_ADDRESS &&
+    buyTokenAddress === WRAPPED_ASSET[chainId]
+  ) {
+    logger.debug(
+      'wrapping native asset',
+      tradeDetails.buyAmount,
+      walletToUse.address,
+      chainId
+    );
+    return wrapNativeAsset(
+      tradeDetails.buyAmount,
+      walletToUse,
+      chainId,
+      transactionParams
+    );
+    // Unwrap Weth
+  } else if (
+    sellTokenAddress === WRAPPED_ASSET[chainId] &&
+    buyTokenAddress === ETH_ADDRESS
+  ) {
+    logger.debug(
+      'unwrapping native asset',
+      tradeDetails.sellAmount,
+      walletToUse.address,
+      chainId
+    );
+    return unwrapNativeAsset(
+      tradeDetails.sellAmount,
+      walletToUse,
+      chainId,
+      transactionParams
+    );
+    // Swap
+  } else {
+    logger.debug(
+      'FILLQUOTE',
+      tradeDetails,
+      transactionParams,
+      walletToUse.address,
+      permit,
+      chainId
+    );
+    return fillQuote(
+      tradeDetails,
+      transactionParams,
+      walletToUse,
+      permit,
+      chainId
+    );
+  }
+};
+
+export const executeCrosschainSwap = async ({
+  chainId,
+  gasLimit,
+  maxFeePerGas,
+  maxPriorityFeePerGas,
+  gasPrice,
+  nonce,
+  tradeDetails,
+  wallet,
+  permit = false,
+  flashbots = false,
+}: {
+  chainId: ChainId;
+  gasLimit: string | number;
+  maxFeePerGas: string;
+  maxPriorityFeePerGas: string;
+  gasPrice: string;
+  nonce?: number;
+  tradeDetails: CrosschainQuote | null;
+  wallet: Wallet | null;
+  permit: boolean;
+  flashbots: boolean;
+}) => {
+  let walletToUse = wallet;
+  const network = ethereumUtils.getNetworkFromChainId(chainId);
+  let provider;
+
+  // Switch to the flashbots provider if enabled
+  if (flashbots && network === Network.mainnet) {
+    logger.debug('flashbots provider being set on mainnet');
+    provider = await getFlashbotsProvider();
+  } else {
+    logger.debug('normal provider being set', network);
+    provider = await getProviderForNetwork(network);
+  }
+
+  if (!walletToUse) {
+    walletToUse = await loadWallet(undefined, true, provider);
+  } else {
+    walletToUse = new Wallet(walletToUse.privateKey, provider);
+  }
+
+  if (!walletToUse || !tradeDetails) return null;
+
+  const { sellTokenAddress, buyTokenAddress } = tradeDetails;
+  const transactionParams = {
+    gasLimit: toHex(gasLimit) || undefined,
+    // In case it's an L2 with legacy gas price like arbitrum
+    gasPrice: gasPrice || undefined,
+    // EIP-1559 like networks
+    maxFeePerGas: maxFeePerGas || undefined,
+    maxPriorityFeePerGas: maxPriorityFeePerGas || undefined,
+    nonce: nonce ? toHex(nonce) : undefined,
+  };
+
+  logger.debug(
+    'FILLCROSSCHAINSWAP',
+    tradeDetails,
+    transactionParams,
+    walletToUse.address,
+    permit,
+    chainId
+  );
+  return fillCrosschainQuote(tradeDetails, transactionParams, walletToUse);
+};
 
 const swap = async (
   wallet: Wallet,
@@ -34,6 +227,7 @@ const swap = async (
     permit,
     chainId,
     requiresApprove,
+    swapType,
   } = parameters as SwapActionParameters;
   const { dispatch } = store;
   const { accountAddress } = store.getState().settings;
@@ -100,7 +294,9 @@ const swap = async (
     };
 
     // @ts-ignore
-    swap = await executeSwap(swapParams);
+    swap = await (swapType === SwapType.crossChain
+      ? executeCrosschainSwap
+      : executeSwap)(swapParams);
     dispatch(
       additionalDataUpdateL2AssetToWatch({
         hash: swap?.hash || '',
